+++ src/App.css
@@ -1,17 +1,23 @@
 .App {
   background-color: grey;
+  overflow: hidden;
 }
 
+/*..........................Primera parte.....................*/
 .HeaderPage {
+  height: 20%;
+  position: relative;
   display: flex;
   justify-content: center;
   align-items: center;
-  height: 20%;
-  box-shadow: -1px -1px 5px 1px black;
+  background-color: #260b2e;
+  box-shadow: 1px 1px 3px 0px black inset;
 }
 
 .TitlePage {
   padding-top: 5px;
+  color: rgb(188, 159, 216);
+  font-size: 29px;
 }
 
 .LogoPremier {
@@ -22,9 +28,9 @@
 
 .List {
   width: 272px;
-  height: 510px;
-  background-color:lightsteelblue;
-  box-shadow: -1px 1px 6px 0px black;
+  height: 528px;
+  background-color: #849595;
+  box-shadow: 0px 1px 3px 0px black;
   float: left;
 }
 
@@ -34,7 +40,7 @@
 }
 
 .TeamsLi {
-  font-size: small;
+  font-size: 17px;
   line-height: 1.3;
   
 }
@@ -47,11 +53,11 @@
 
 .InfoTeam {
   position: relative;
-  background-color: darkslategray;
-  box-shadow: -1px -1px 5px 1px black;
+  background-color: #849595;
   width: 80%;
-  height: 80%;
+  height: 528px;
   margin-left: 274px;
+  box-shadow: 1px 1px 3px 1px black;
 }
 
 .DivTeamLogo {
@@ -102,15 +108,15 @@
 
 .DivStadiumLogo {
   width: 356px;
-  height: 253px;
+  height: 272px;
   box-shadow: -1px -1px 5px 1px black;
 }
 
 .StadiumLogo {
   width: 357px;
-  height: 253px;
+  height: 271px;
 }
-/*------- arreglar----------*/
+/*------- Jugadores ----------*/
 
 .DivPlayersTeam {
   position: absolute;
@@ -193,4 +199,124 @@
 .PlayersFeaturesLi {
   margin-left: 1px;
   margin-bottom: 2px;
+}
+
+/* ..................Segunda parte.................... */
+
+.PageMenu {
+  position: absolute;
+  width: 243px;
+  height: 18px;
+  margin-top: 49px;
+  margin-left: 1178px;
+}
+
+a {
+  margin-left: 1px;
+  margin-right: 2px;
+  background-color: #102638;
+  color: rgb(131 129 129);
+  padding: 5px;
+  text-decoration: none;
+}
+
+a:hover {
+  background-color: rgb(3, 28, 39);
+}
+
+.BodyPos {
+ width: 100%;
+ height: 878px;
+ position: relative;
+ background-color: #849595;
+ box-shadow: 1px 1px 4px 1px black inset;
+}
+
+.DivTable {
+  width: 80%;
+  height:80%;
+  position: absolute;
+  margin-top: 1px;
+  margin-left: 300px;
+}
+
+table {
+  width: 70%;
+  margin-top: 105px;
+  font-family: "Lucida Sans Unicode", "Lucida Grande", Sans-Serif;
+  font-size: 15px;
+  border-collapse: collapse;
+}
+th { 
+  font-size: 12px;
+  font-weight: normal;
+  text-align: center;  
+  padding: 7px;     
+  background: #090a0e;
+  border-bottom: 1px solid #813b3b; 
+  color: #0f2e6c;
+}
+
+td {  
+  text-align: center;
+  padding: 1px;    
+  background: #0c0e18;   
+  border-bottom: 1px solid #813b3b;
+  color: #aeaec1; 
+  border-top: 1px solid transparent;
+}
+
+thead th:nth-child(1) {
+  width: 65%;
+}
+
+.ClubRank {
+  float: left;
+  width: 30px;
+  height: 25px;
+  padding-left: 8px;
+  padding-top: 9px;
+  font-size: small;
+}
+
+.ClubLogo {
+  float: left;
+  width: 42px;
+  height: 25px;
+  padding-top: 5px;
+}
+
+.TableTeamLogo {
+  width: 25px;
+  height: 25px;
+}
+
+.ClubName {
+  float: left;
+  width: 180px;
+  height: 23px;
+  padding-left: 2px;
+  padding-top: 9px;
+  font-size: 13px;
+  text-align: left;
+}
+
+.DivTitlePos {
+  width: 29%;
+  height: 35px;
+  position: absolute;
+  margin-left: 303px;
+  margin-top: 59px;
+}
+
+.TitlePositions {
+  
+  width: 88%;
+  height: 20px;
+  font-size: 18px;
+}
+
+.ClubAllPoints {
+  width: 32px;
+  height: 12px;
 }
