+++ src/index.css
@@ -5,8 +5,7 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background-color: grey;
-  overflow: hidden;
+  overflow-x: hidden;
 }
 
 code {
